modified resolution scaling to support upscaling

+++ BAF_engine/include/graphics.h
@@ -65,7 +65,7 @@ class Graphics
     World *world;
     Object *renderTarget;
 
-    int resScale;
+    float resScale;
 
     int width;
     int height;
